requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
whitenoise==5.3.0
psycopg2-binary